Feat (src/all): add collision ez sinan

+++ src/utils/do_aabb.c
@@ -47,11 +47,11 @@ bool do_aabb_sprites_rectangle(sfSprite *sp1, sfRectangleShape *sp2)
     sfIntRect r1 = sfSprite_getTextureRect(sp1);
     sfVector2f r2 = sfRectangleShape_getSize(sp2);
 
-    if (r2.y < 0)
-        r2.y *= -1;
-    if (pos1.x < pos2.x + r2.x &&
+    if (r2.y < (float)0)
+        r2.y *= (float)-1;
+    if (pos1.x < pos2.x + (float)r2.x &&
     pos1.x + r1.width > pos2.x &&
-    pos1.y < pos2.y + r2.y &&
+    pos1.y < pos2.y + (float)r2.y &&
     pos1.y + r1.height > pos2.y) {
         return true;
     }
